pyproject.toml
[project]
name = "cordis"
version = "0.1.0"
description = "A file and artifact management service built for large-scale objects."
authors = [
    { name = "nhanphanvan99", email = "nhanphanvan99@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic (>=1.14.1,<2.0.0)",
    "aiosqlite (>=0.20.0,<0.21.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "typer (>=0.15.1,<0.16.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "tenacity (>=9.1.4,<10.0.0)",
]

[project.scripts]
cordis = "cordis.cli.main:app"
cordis-server = "cordis.backend.__main__:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "cordis" }]

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.5"
mypy = "^1.15.0"
pylint = "^3.3.4"
black = "^26.3.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^1.3.0"

[tool.ruff]
target-version = "py310"
line-length = 120
output-format = "concise"
show-fixes = true
unsafe-fixes = false
preview = true

[tool.ruff.lint]
select = [
    "E",
    "F",
    "B",
    "UP",
    "I",
    "Q",
    "W",
]
unfixable = ["B"]

[tool.ruff.format]
docstring-code-format = true
quote-style = "double"

[tool.pylint]
disable = [
    "too-many-instance-attributes",
    "too-many-positional-arguments",
    "too-few-public-methods",
    "too-many-return-statements",
    "too-many-locals",
    "too-many-arguments",
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "import-outside-toplevel",
    "cyclic-import",
    "redefined-builtin",
    "broad-exception-caught",
    "broad-exception-raised",
    "super-init-not-called",
    "duplicate-code",
    "fixme",
    "logging-fstring-interpolation",
    "arguments-differ",
    "too-many-statements",
    "too-many-public-methods",
    "invalid-name",
]
persistent = false
reports = false
max-branches = 20
output-format = "colorized"
max-line-length = 120

[tool.mypy]
no_implicit_optional = false
ignore_missing_imports = true
check_untyped_defs = true
warn_return_any = true
warn_no_return = false
warn_unreachable = true
strict = true
strict_optional = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
